+++ libraries/data/numerical_recipes.h
@@ -99,12 +99,12 @@ inline size_t fact(int num)
 
 }
 
-inline double binom(int n, int k)
+inline size_t binom(int n, int k)
 {
     size_t factor=1;
-    for (size_t i = n; i > (n-k); --i)
+    for (int i = n; i > (n-k); --i)
         factor *= i;
-    return double(factor)/double(fact(k));
+    return factor/fact(k);
 }
 
 // Singular value descomposition of matrix a (numerical recipes, chapter 2-6 for details)
